postgres/requirements.txt
psycopg2-binary==2.9.3
pandas==2.1.1 
numpy==1.25.0